Moved DmapError definition into its own file, error.rs.
